Directive to include code without license header
This is implemented in the `no-license` directive, which has the
exact same usage and options as `literalinclude`. Only difference
is the `lines` options is set to start after the license header
and to end at the last line of the file.

The end of the license header is detected by searching for the
line that contains the SPDX identifier. Therefore this directive
assumes that the license header contains the SPDX identifier.

Signed-off-by: Daniel Lim Wee Soong <weesoong.lim@gmail.com>